+++ src/liveconfigs/admin.py
@@ -1,7 +1,7 @@
-import datetime as dt
 import json
 
 from django.contrib import admin
+from django.utils import timezone
 from import_export import formats, resources
 from import_export.admin import ImportExportModelAdmin
 from import_export.fields import Field
@@ -9,7 +9,7 @@ from tablib import Dataset
 
 from .filters import TagsListFilter
 from .forms import ConfigRowForm
-from .models import ConfigRow
+from .models import ConfigRow, HistoryEvent
 from .utils import get_excluded_rows
 
 
@@ -48,24 +48,23 @@ class ConfigRowAdmin(ImportExportModelAdmin):
     form = ConfigRowForm
     formats = [formats.base_formats.JSON]
     resource_class = ConfigRowResource
-    list_display = ('name', 'value', 'description', 'topic',
-                    'tags', 'last_read', 'last_set')
+    list_display = ('name', 'value', 'default_value', 'description', 'topic', 'tags', 'last_read', 'last_set')
     list_filter = ("topic", TagsListFilter,)
-    readonly_fields = ('name', 'description', 'topic',
-                       'tags', 'last_read', 'last_set')
+    readonly_fields = ('name', 'description', 'topic', 'tags', 'last_read', 'last_set', 'default_value')
     search_fields = ('name', 'description', 'topic', 'tags')
 
-    def get_export_queryset(self, request):
-        qs = ConfigRow.objects.all()
-        excluded_config_rows = get_excluded_rows()
-        return qs.exclude(name__in=excluded_config_rows)
-
     def save_model(self, request, obj, form, change):
-        """
-        Given a model instance save it to the database.
-        """
-        obj.last_set = dt.datetime.now(tz=dt.timezone.utc)
-        obj.save()
+        user = request.user
+        super().save_model(request, obj, form, change)
+        HistoryEvent.objects.create(name=obj.name, value=obj.value, edit_at=timezone.now(), edit_by=user)
+
+
+@admin.register(HistoryEvent)
+class HistoryEventAdmin(admin.ModelAdmin):
+    list_display = ("name", "value", "edit_at", "edit_by")
+    readonly_fields = list_display
+    search_fields = ("name",)
+    list_filter = ("name", "edit_at")
 
 
 admin.site.register(ConfigRow, ConfigRowAdmin)